Visualization System/requirements.txt
Flask==3.1.2
Flask_Cors==5.0.0
ftfy==6.3.1
matplotlib==3.7.5
numpy==1.23.2
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
packaging==25.0
Pillow==11.3.0
pybind11==2.13.6
pytest==8.4.1
regex==2024.11.6
Requests==2.32.5
scikit_learn==1.3.2
scikit_learn_extra==0.3.0
scipy==1.16.1
segment_anything==1.0
setuptools==75.1.0
Shapely==2.1.1
skimage==0.0
submitit==1.5.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
